requirements.txt
streamlit==1.31.0
langchain==0.3.14
langchain-community==0.3.14
langchain-openai==0.2.14
langchain-text-splitters==0.3.4
openai==1.58.1
chromadb==0.4.22
python-dotenv==1.0.0
tiktoken==0.8.0
